requirements.txt
mlflow==1.24.0
numpy==1.19.5
pandas==1.4.1
pandas_profiling==3.1.0
python-dotenv==0.19.2
scikit_learn==1.0.2
setuptools==60.5.0
xgboost==1.5.2
